fix: removing the ability to delete an administrator account

+++ src/main/java/app/web/MyProfileController.java
@@ -1,11 +1,13 @@
 package app.web;
 
+import app.exception.AccessDeniedException;
 import app.exception.EmailAlreadyExistsException;
 import app.exception.NotAuthenticatedException;
 import app.mapper.user.UserMapper;
 import app.model.dto.user.MyProfileUpdateRequest;
 import app.model.dto.user.UserSession;
 import app.model.entity.user.Region;
+import app.model.entity.user.UserRole;
 import app.repository.user.UserRepository;
 import app.service.user.UserService;
 import app.service.user.UserSessionService;
@@ -57,6 +59,7 @@ public class MyProfileController {
         }
 
         model.addAttribute("username", userSession.getUsername());
+        model.addAttribute("isAdmin", userSession.getRole() == UserRole.ADMIN);
         return "my-profile";
     }
 
@@ -68,6 +71,7 @@ public class MyProfileController {
                                   RedirectAttributes redirectAttributes) {
         UserSession userSession = requireUserSession(session);
         model.addAttribute("username", userSession.getUsername());
+        model.addAttribute("isAdmin", userSession.getRole() == UserRole.ADMIN);
 
         validatePasswordFields(request, bindingResult);
 
@@ -91,6 +95,11 @@ public class MyProfileController {
                                 HttpServletRequest request,
                                 HttpServletResponse response) {
         UserSession userSession = requireUserSession(session);
+
+        if (userSession.getRole() == UserRole.ADMIN) {
+            throw new AccessDeniedException();
+        }
+
         userService.deleteAccount(userSession.getId());
 
         new SecurityContextLogoutHandler().logout(
